feat: refactor BetAmount component and add new sections for mobile and desktop views; integrate use-debounce for improved performance;
