+++ package.json
@@ -43,6 +43,7 @@
 		"react-virtualized": "^9.22.6",
 		"tailwindcss": "^4.1.3",
 		"tw-animate-css": "^1.2.5",
+		"use-debounce": "^10.0.5",
 		"viem": "^2.26.2",
 		"wagmi": "^2.14.16"
 	},
