--- src/components/VersionValidation.tsx
@@ -1,32 +1,0 @@
-import logger from '@/src/config/logger.ts';
-import { Button, toast } from '@betfinio/components/ui';
-import { useLatestVersion } from 'betfinio_context/lib/query';
-import { type FC, useEffect } from 'react';
-
-interface VersionValidationProps {
-	repository: string;
-	branch: string;
-	current: string;
-}
-export const VersionValidation: FC<VersionValidationProps> = ({ branch, repository, current }) => {
-	const { data: version } = useLatestVersion(repository, branch);
-	useEffect(() => {
-		logger.warn('Latest version:', version);
-		logger.warn('Deployed version:', current);
-		if (!version || !current || current === 'undefined') return;
-		if (version.toLowerCase() !== current.toLowerCase()) {
-			logger.warn('New version available!');
-			toast.info('New version available!', {
-				action: (
-					<Button size={'sm'} shape={'pill'} onClick={handleRefresh}>
-						Update
-					</Button>
-				),
-			});
-		}
-	}, [version, current, repository, branch]);
-	const handleRefresh = () => {
-		window.location.reload();
-	};
-	return null;
-};
